expanded agent, updated method to MC intead of MVC
